frontend/requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
boto3==1.21.18
botocore==1.24.18
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Django==4.0.3
django-fernet-fields==0.6
django-recaptcha==3.0.0
django-widget-tweaks==1.4.12
dropbox==11.28.0
idna==3.3
jmespath==0.10.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
paramiko==2.9.2
pathspec==0.9.0
platformdirs==2.5.1
ply==3.11
psutil==5.9.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stone==3.3.1
tomli==2.0.1
urllib3==1.26.8
xmltodict==0.12.0
